Add class "RTCRtpTransceiver"
