+++ Runtime/Scripts/RTCRtpTransceiver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bae45dcf03c88444f8abc22f9d0e257e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
