+++ Runtime/Scripts/WebRTC.cs
@@ -13,14 +13,6 @@ namespace Unity.WebRTC
         Hardware = 1
     }
 
-    public enum RTCRtpTransceiverDirection
-    {
-        SendRecv,
-        SendOnly,
-        RecvOnly,
-        Inactive
-    }
-
     public struct RTCIceCandidate​
     {
         [MarshalAs(UnmanagedType.LPStr)]
@@ -67,75 +59,6 @@ namespace Unity.WebRTC
         internal RTCRtpSender(IntPtr ptr)
         {
             self = ptr;
-        }
-    }
-
-    public class RTCRtpTransceiver
-    {
-        internal IntPtr self;
-
-        internal RTCRtpTransceiver(IntPtr ptr)
-        {
-            self = ptr;
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public RTCRtpTransceiverDirection CurrentDirection
-        {
-            get
-            {
-                var direction = RTCRtpTransceiverDirection.RecvOnly;
-                if (NativeMethods.TransceiverGetCurrentDirection(self, ref direction))
-                {
-                    return direction;
-                }
-                throw new InvalidOperationException("Transceiver is not running");
-            }
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public RTCRtpReceiver Receiver
-        {
-            get { return new RTCRtpReceiver(NativeMethods.TransceiverGetReceiver(self));  }
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public RTCRtpSender Sender
-        {
-            get { return new RTCRtpSender(NativeMethods.TransceiverGetSender(self)); }
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="direction"></param>
-        public void SetDirection(RTCRtpTransceiverDirection direction)
-        {
-            // TODO::
-            throw new NotImplementedException();
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public struct RTCRtpCodecCapability
-        {
-        }
-
-        public void SetCodecPreferences(RTCRtpCodecCapability[] capabilities)
-        {
-
-        }
-
-        public void Stop()
-        {
-            NativeMethods.TransceiverStop(self);
         }
     }
